+++ etl-service/src/main/java/com/sitefilm/etl/dto/core/movie/MovieIdDto.java
@@ -1,5 +1,5 @@
 package com.sitefilm.etl.dto.core.movie;
 
 public record MovieIdDto(
-        Integer id
+        Long id
 ){}
